pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "skillgen"
version = "0.1.0"
description = "Generate Agent Skills from llms.txt"
readme = "README.md"
requires-python = ">=3.9"
license = {file = "LICENSE"}
authors = [{name = "Mihir Srivastava"}]
keywords = ["llms.txt", "agent", "skills", "documentation", "generator"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent"
]
dependencies = [
  "requests>=2.31.0",
  "PyYAML>=6.0",
  "trafilatura>=1.8.0",
  "beautifulsoup4>=4.12.0",
  "mcp>=1.0.0"
]

[project.urls]
Homepage = "https://github.com/mihir-s-05/skillgen"
Repository = "https://github.com/mihir-s-05/skillgen"
Issues = "https://github.com/mihir-s-05/skillgen/issues"

[project.optional-dependencies]
# Optional fallback converter
extras = ["markdownify>=0.11.6"]

test = ["pytest>=7.4"]

[project.scripts]
skillgen = "skillgen.cli:main"
skillgen-mcp = "skillgen.mcp_server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["skillgen*"]
